Nacho/cleanup 3rdparty cmake (#129)
* Cleanup Sophus.cmake build script

* Cleanup eigen target

We now run cmake and check the compiler actually supports eigen. Before
it was a raw copy-paste of the code

* Attempt to improve the cmake usage

* add tsl to the same idea

* Fix lol

* Simplify

* typo

* Use own fork again

* cleanup script

* Almost there with sophus

* Fix sophus target

* Bump min required cmake version

This is to make sure that FetchContent_Declare has the SYSTEM flag
avaialbe. This improves the superbuild usage in general

* Bump bit more the cmake min requirement

* Be less strict for developers who actually have installed the
dependencies

* Bump cmake version in CI

* Ignore this

* Attempt to fix windows build

* ROS2 now also requires to have cmake 3.25

This is the main disadvantage of this whole change

* tbb release mode to fix windows build