Support Pydantic V2 (#265)
* Support Pydantic V2

This updates the code to be compatible with Pydantic V2. We also
introduce a smoke test for importing `kiss_icp`. It's still just a smoke
test, but does a bit more than only calling `kiss_icp_pipeline
--version`. The latter would also pass with Pydantic V2 being installed,
while the import would fail.

Co-authored-by: pnikutta <150661207+pnikutta@users.noreply.github.com>

* Add make target

* Remove debug msg

The config file is in the output directory anyways

* Remove unused import

---------

Co-authored-by: pnikutta <150661207+pnikutta@users.noreply.github.com>
Co-authored-by: Ignacio Vizzo <ignacio@dexory.com>